Fixed tests for consistency
- Moved courses-index view test to tests/views
- Moved test_models to tests/models
- Update module docstring in tests
- Put each model test class into it's own test file
- Removed sanity check test as it is not needed
