+++ tests/models/test_course_daily_metrics_model.py
@@ -1,5 +1,5 @@
-'''
-Test models declared in edX Figures
+'''Tests edx-figures CourseDailyMetrics model
+
 '''
 
 import datetime
@@ -7,17 +7,14 @@ import pytest
 
 from django.db.utils import IntegrityError
 
-from edx_figures.models import CourseDailyMetrics, SiteDailyMetrics
+from edx_figures.models import CourseDailyMetrics
 
-from .factories import (
-    CourseDailyMetricsFactory,
-    SiteDailyMetricsFactory,
-    )
+from tests.factories import CourseDailyMetricsFactory
 
 
 @pytest.mark.django_db
 class TestCourseDailyMetrics(object):
-    '''Unit tests for the CourseDailyMetrics model
+    '''"Tests the CourseDailyMetrics model
 
     Focuses on testing CourseDailyMetrics fields and methods
 
@@ -26,7 +23,7 @@ class TestCourseDailyMetrics(object):
     '''
     @pytest.fixture(autouse=True)
     def setup(self, db):
-        '''Placeholder for
+        '''Placeholder for test setup
         '''
         pass
 
@@ -92,52 +89,3 @@ class TestCourseDailyMetrics(object):
         and that the course_id is also not an empty string
         '''
         pass
-
-
-@pytest.mark.django_db
-class TestSiteDailyMetrics(object):
-    '''Unit tests for  the SiteDailyMetrics model
-
-    Focus on testing SiteDailyMetrics methods and fields
-    '''
-
-    @pytest.fixture(autouse=True)
-    def setup(self, db):
-        '''
-
-        '''
-        self.site_daily_metrics = [
-            SiteDailyMetricsFactory()
-        ]
-
-    def test_foo(self):
-        '''
-        Assert that SiteDailyMetricsFactory works by checking the object
-        created in this class's ``setup`` method.
-
-        '''
-        assert SiteDailyMetrics.objects.count() == 1
-
-
-    @pytest.mark.parametrize('rec', [
-        dict(
-            date_for=datetime.date(2018,02,02),
-            defaults=dict(
-                cumulative_active_user_count=11,
-                total_user_count=1,
-                course_count=1,
-                total_enrollment_count=1
-            ),
-        ),
-    ])
-    def test_create(self, rec):
-        '''Sanity check we can create the SiteDailyMetrics model
-
-        Create a second instance the way we'll do it in the production code.
-        Assert this is correct
-        '''
-
-        site_metrics, created = SiteDailyMetrics.objects.get_or_create(**rec)
-
-        assert created and site_metrics
-
