--- tests/test_sanity_check.py
@@ -1,10 +1,0 @@
-'''
-This test file is to sanity check that the basic test framework works
-'''
-
-def test_sanity_check():
-    '''
-    Just asserts True to make sure that this test succesfully runs.
-    If it doesn't then there is a problem with the project setup
-    '''
-    assert True
