+++ tests/views/test_site_daily_metrics_view.py
@@ -1,4 +1,4 @@
-'''Unit tests for the site daily metrics view
+'''Tests edx-figures SiteDailyMetricsViewSet class
 
 '''
